Create and use a wrapper script to run gofmt
The wrapper script emits the original contents in case of a syntax error, for
now, and automatically injects/removes a 'package ' line, which makes it also
suitable as equalprg (for the '=' operation in normal mode).  Use this wrapper
script for the on-write gofmt run, and set equalprg to it as well.

Signed-off-by: Chris Larson <chris_larson@mentor.com>